This update adds additional support for file.  Files can now be opened closed, read, written in bytes by position (block reads and writes) and piped to other files during idle periods (for large files).
check_size.sh has a modification for OSX that allows the same
functionality as on other unix like environments.  Since DU does not
function the same, gdu must be loaded on the system (through macports or
some other channel). The modification is simply to use gdu instead of
du.

This commit also includes a QT creator project file and associated
settings.