Hopefully fix repeated GET problem
Previously, if we issued http GET more than once every 30 seconds, the second GET would timeout.  It seems this is because the code was using the same fixed port for src and dest - both were 80.  Probably this caused a problem if the second GET happened before the first had closed