WIP: a better dockerfile
Hi, there is still a lot of improvements to be made with the Dockerfile. Currently, it is not cache-friendly and container image rebuild takes a lot of time. We should run the `provision.sh` script first and then copy all the source code into the container. This way docker will only re-run the last steps of the Dockerfile and save a lot of time. 

I tried to do this, and also I changed the last line to be
```Dockerfile
CMD "make"
```
instead of
```Dockerfile
CMD ["bash", "-c", "source scripts/provision.sh ALL && make"]
```

and when I try to run the container I get this error:
```
make: arm-none-eabi-gcc: Command not found
```

Why is that happening? `arm-none-eabi-gcc` should already be installed during the build of a docker image, isn't it?