Update this reference in call/apply/bind types
`this` as the first argument means the function is callable as in
`obj.func()`, and that we've specified the type of `obj`, rather
than the type of the first argument.

We can rename `this` to `thisArg` to avoid this behaviour, meaning
we now have a "real" first argument, that will be the `this` passed
to the function we've specified.

This change also restricts `apply`'s second argument to be an
array-like object.