Fix `if (0);"test"||fail()` 2v19 regression - strings after if(0) didn't get interpreted Ensure eval("1;;;")==1 - eval("1;")==1 before, but not eval("1;;")
We hit this with minification in the AT library... If there's a string right after an if(0) it
doesn't get parsed!

This happens because since 2v19ish we're no longer storing string contents in RAM if
we're not supposed to be executing. But we need to be sure we don't Lex too far ahead
with execute=false set