Bangle_setUI_Q3: only add custom btn handler
This is maybe a quick fix. There could maybe be a bigger refactor to
handle btn watches in the same way as other handlers are added, where
they aren't added until we know which on we want to set (By
writing/overwriting e.g. `Bangle.touchHandler` and only actually set it
with `Bangle.on` at the end of the script).