Bangle_setUI_Q3: Make custom btn edge configurable
`btn` can be set like before, so backwards compatible I think.
But it can now also be an object on form
{fn: ()={}, edge: "rising/falling/both"}.
@gfwilliams has suggested another solution using callbacks in
https://github.com/espruino/BangleApps/pull/3452#discussion_r1642419463
The solution I opted for here for now follows the same coding
style of how the `type`/`mode` parameter of setUI can be
either a string or an object.